+++ lib/FirelinkCore/include/FirelinkCore/GameType.h
@@ -4,15 +4,19 @@
 
 namespace Firelink
 {
+    //! @brief Game type enum for supported FromSoftware games.
+    //! @details Must be kept in sync with `soulstruct` enum.
     enum class GameType : std::uint8_t
     {
-        DemonsSouls = 0,
+        DemonsSouls = 0, // PS5 Remake will never be supported
         DarkSoulsPTDE = 1,
         DarkSoulsDSR = 2,
-        Bloodborne = 3,
-        DarkSouls3 = 4,
-        Sekiro = 5,
-        EldenRing = 6,
+        DarkSouls2 = 3, // not really supported
+        DarkSouls2SOTFS = 4, // not really supported
+        Bloodborne = 5,
+        DarkSouls3 = 6, // barely supported
+        Sekiro = 7, // barely supported
+        EldenRing = 8,
     };
 
     ///! @brief Returns true if `game` generally uses DCX for Binder compression.